# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Canarytokens API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TokenType = (typeof TokenType)[keyof typeof TokenType];

export const TokenType = {
  web: "web",
  dns: "dns",
  email: "email",
  pdf: "pdf",
  word: "word",
  qr_code: "qr_code",
  image: "image",
} as const;

export interface Token {
  id: string;
  type: TokenType;
  name: string;
  memo: string;
  token: string;
  triggerUrl: string;
  alertEmail?: string | null;
  triggered: boolean;
  triggerCount: number;
  createdAt: string;
  lastTriggeredAt?: string | null;
}

export interface CreateTokenRequest {
  type: TokenType;
  name: string;
  memo: string;
  alertEmail?: string | null;
}

export interface TokenListResponse {
  tokens: Token[];
  total: number;
  page: number;
  limit: number;
}

export interface Alert {
  id: string;
  tokenId: string;
  ipAddress?: string | null;
  userAgent?: string | null;
  referer?: string | null;
  geo?: string | null;
  triggeredAt: string;
}

export interface AlertListResponse {
  alerts: Alert[];
  total: number;
  page: number;
  limit: number;
}

export interface StatsResponse {
  totalTokens: number;
  triggeredTokens: number;
  totalAlerts: number;
  recentAlerts: Alert[];
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export type ListTokensParams = {
  page?: number;
  limit?: number;
};

export type ListTokenAlertsParams = {
  page?: number;
  limit?: number;
};

export type TriggerTokenParams = {
  ref?: string;
};