# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Canarytokens API
 * OpenAPI spec version: 0.1.0
 */

export interface Alert {
  id: string;
  tokenId: string;
  ipAddress?: string | null;
  userAgent?: string | null;
  referer?: string | null;
  geo?: string | null;
  triggeredAt: Date;
}