# lib/api-zod/src/generated/types/alertListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Canarytokens API
 * OpenAPI spec version: 0.1.0
 */
import type { Alert } from "./alert";

export interface AlertListResponse {
  alerts: Alert[];
  total: number;
  page: number;
  limit: number;
}