# lib/api-zod/src/generated/types/createTokenRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Canarytokens API
 * OpenAPI spec version: 0.1.0
 */
import type { TokenType } from "./tokenType";

export interface CreateTokenRequest {
  type: TokenType;
  name: string;
  memo: string;
  alertEmail?: string | null;
}