# lib/api-zod/src/generated/types/statsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Canarytokens API
 * OpenAPI spec version: 0.1.0
 */
import type { Alert } from "./alert";

export interface StatsResponse {
  totalTokens: number;
  triggeredTokens: number;
  totalAlerts: number;
  recentAlerts: Alert[];
}