# lib/api-zod/src/generated/types/token.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Canarytokens API
 * OpenAPI spec version: 0.1.0
 */
import type { TokenType } from "./tokenType";

export interface Token {
  id: string;
  type: TokenType;
  name: string;
  memo: string;
  token: string;
  triggerUrl: string;
  alertEmail?: string | null;
  triggered: boolean;
  triggerCount: number;
  createdAt: Date;
  lastTriggeredAt?: Date | null;
}