# lib/api-zod/src/generated/types/tokenListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Canarytokens API
 * OpenAPI spec version: 0.1.0
 */
import type { Token } from "./token";

export interface TokenListResponse {
  tokens: Token[];
  total: number;
  page: number;
  limit: number;
}