# lib/api-zod/src/generated/types/tokenType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Canarytokens API
 * OpenAPI spec version: 0.1.0
 */

export type TokenType = (typeof TokenType)[keyof typeof TokenType];

export const TokenType = {
  web: "web",
  dns: "dns",
  email: "email",
  pdf: "pdf",
  word: "word",
  qr_code: "qr_code",
  image: "image",
} as const;